requirements.txt
altgraph==0.17
cachetools==4.0.0
canvasapi==0.15.0
certifi==2019.11.28
chardet==3.0.4
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.17.0
idna==2.8
oauthlib==3.1.0
pefile==2019.4.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyInstaller==3.6
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
